Make BeatmapConverter more legible
