+++ osu.Game.Rulesets.Sentakki/TestSceneSentakkiPlayer-150ea300-170f-4435-9615-a0fbebadd7e0/game.ini
@@ -1,0 +1,59 @@
+Ruleset = 0
+Skin = 0
+BeatmapDetailTab = Details
+ShowConvertedBeatmaps = True
+DisplayStarsMinimum = 0.0
+DisplayStarsMaximum = 10.1
+SongSelectGroupingMode = All
+SongSelectSortingMode = Title
+RandomSelectAlgorithm = RandomPermutation
+ChatDisplayHeight = 0.4
+Username = 
+Token = 
+SavePassword = False
+SaveUsername = True
+ExternalLinkWarning = True
+PreferNoVideo = False
+VolumeInactive = 0.25
+MenuVoice = True
+MenuMusic = True
+AudioOffset = 0.0
+MenuCursorSize = 1.0
+GameplayCursorSize = 1.0
+AutoCursorSize = False
+MouseDisableButtons = False
+MouseDisableWheel = False
+ShowFpsDisplay = False
+ShowStoryboard = True
+BeatmapSkins = True
+BeatmapHitsounds = True
+CursorRotation = True
+MenuParallax = True
+DimLevel = 1.0
+BlurLevel = 0.0
+LightenDuringBreaks = True
+HitLighting = True
+ShowInterface = True
+ShowProgressGraph = True
+ShowHealthDisplayWhenCantFail = True
+FadePlayfieldWhenHealthLow = True
+KeyOverlay = False
+PositionalHitSounds = True
+ScoreMeter = HitErrorBoth
+FloatingComments = False
+ScoreDisplayMode = Standardised
+IncreaseFirstObjectVisibility = True
+ReleaseStream = Lazer
+Version = 
+ScreenshotFormat = Jpg
+ScreenshotCaptureMenuCursor = False
+SongSelectRightMouseScroll = False
+Scaling = Off
+ScalingSizeX = 0.8
+ScalingSizeY = 0.8
+ScalingPositionX = 0.5
+ScalingPositionY = 0.5
+UIScale = 1.0
+UIHoldActivationDelay = 200.0
+IntroSequence = Triangles
+MenuBackgroundSource = Skin
